requirements.txt
# Core
python-binance==1.0.19
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Database
sqlalchemy==2.0.23

# Scheduling
schedule==1.2.0

# ML & Prediction
scikit-learn==1.3.0
xgboost==2.0.3
prophet==1.1.5

# Utilities
colorlog==6.7.0
tqdm==4.66.1

requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
schedule==1.2.0
colorlog==6.7.0
tqdm==4.66.1
scikit-learn==1.3.0
lightgbm==4.1.0